write loosely coupled variant of check-membership
