+++ pallets/check-membership/Cargo.toml
@@ -33,4 +33,5 @@ frame-system = { version = '2.0.0-dev', git = 'https://github.com/paritytech/sub
 sp-runtime = { version = '2.0.0-dev', git = 'https://github.com/paritytech/substrate', rev = '45b9f0a9cbf901abaa9f1fca5fe8baeed029133d', default-features = false }
 sp-core = { version = '2.0.0-dev', git = 'https://github.com/paritytech/substrate', rev = '45b9f0a9cbf901abaa9f1fca5fe8baeed029133d', default-features = false }
 sp-std = { version = '2.0.0-dev', git = 'https://github.com/paritytech/substrate', rev = '45b9f0a9cbf901abaa9f1fca5fe8baeed029133d', default-features = false }
+account-set = { path = '../../traits/account-set' }
 vec-set = { path = '../vec-set' }
